--- a/Math/NumberTheory/Moduli/Equations.hs
+++ b/Math/NumberTheory/Moduli/Equations.hs
@@ -14,6 +14,7 @@
 
 module Math.NumberTheory.Moduli.Equations
   ( solveLinear
+  , solveLinear'
   , solveQuadratic
   ) where