… Math/NumberTheory/Moduli/PrimitiveRoot.hs
… arithmoi.cabal
… benchmark/Bench.hs
… benchmark/Math/NumberTheory/DiscreteLogarithmBench.hs
… test-suite/Math/NumberTheory/Moduli/DiscreteLogarithmTests.hs
… test-suite/Math/NumberTheory/Moduli/PrimitiveRootTests.hs
… test-suite/Test.hs