Discrete logarithms and Dirichlet characters. Resolves #88. In roughly this order, this is my plan for full efficient implementation of both.

- [x] Added `MultMod` type to represent elements of the multiplicative group mod m, with Monoid instance. Added `PrimitiveRoot` type to represent primitive roots of the (cyclic) multiplicative group.
Note these two types together make `discreteLogarithm` a total function. Partial constructors have been added and deconstructors. (I would like feedback on the name `MultMod`, I feel a better name is possible.)
- [x] Amended primitive root code, testing, benching with the new type structures.
- [x] Very simple discrete log code, with test and bench.
- [x] Baby-step giant-step method for discrete logs.
- [x] Additional testing and benching for discrete logs, with prime power cases included also.
- [x] Refactor discrete logs to be efficient on prime powers, using Bach's method described [here](https://math.stackexchange.com/questions/1863037/discrete-logarithm-modulo-powers-of-a-small-prime) and [here](https://www2.eecs.berkeley.edu/Pubs/TechRpts/1984/CSD-84-186.pdf). (Also use the canonical isomorphism for double prime powers).
- [x] Implement Pollard's rho algorithm for discrete logs with prime moduli (#129 would be helpful for this). Notably, Pollard's rho is a Las Vegas algorithm (which is why @Taneb's attempt in #110 did not always return success) so may need to run multiple times - experiment with starting inputs.