@Bodigrim 2018-08-23
> Implement Pollard's rho algorithm for discrete logs with prime moduli (#129 would be helpful for this)

I pushed `solveLinear` into `master`.