@b-mehta 2018-08-30
The Bach reduction gave a large speedup for prime powers: the case `2^n = 7 mod 3^20` went from around 41ms to 3.5µs and `3^n = 17 mod 5^16` went from 549ms to 3.8µs.  Range benchmarks did not change significantly, which is expected since prime powers are rare compared to primes in general.