@b-mehta 2018-08-31
I've added Pollard's rho but on benchmarking, it doesn't have the speedup I expected. Specifically, for moduli below around 10^8 it is slower than the baby-step giant-step method, but works faster for large moduli. For cases near 10^6, the BSGS method took 27.27ms while Pollard's rho took 43.15ms while for cases near 10^10, BSGS took 8.61s and Pollard's rho took 4.06s. 

While it's not surprising that Pollard's rho is eventually faster, I'm surprised the cutoff is so high. Is there some obvious deficiency in my rho algorithm that makes it perform badly on small inputs? @Bodigrim

(Note the rho implementation isn't perfect right now - it currently only works _probabilistically_ but the probability of error is so small that it succeeds on all the test cases and benchmark cases. The functions `starter` and `begin` are part of the fix for this, but right now they appear redundant, and when failure occurs, the error will be from calling `head` on an empty list in line 79. I'll fix in later commits.)