@Bodigrim 2018-08-31
Sorry, it is my fault; maybe wrong compilation parameters. I can confirm: your code does not allocate triples as well.

> Your code doesn't succeed when I test it - there are many missing rem calls which cause it to fail.

What do you mean? It seems passing tests in my environment.