@Bodigrim 2018-08-31
Got it. Prior to GHC 8.4 both `powModInteger` and `recipModInteger` were broken for a negative first argument (https://ghc.haskell.org/trac/ghc/ticket/14085).

That is why after my modification `solveLinear'` started to return weird results in GHC 8.2 and led to an empty `filter check $ begin (starter 0 0)`. I have pushed fixes for `solveLinear'` into `master` and now my implementation works fine for all supported range of GHCs (https://travis-ci.org/Bodigrim/arithmoi/builds/423235835). 20-30% improvement in benchmarks is observed, same to reported above. I suspect it is purely due to changes in `step` function.

Could you please rebase your branch upon the latest `master`?