@b-mehta 2018-09-01
It passes tests but I'm not fully convinced of correctness - you've removed the `rem` after `ai+1` and the like, but the remainder still should be taken after these additions, since they are likely to get larger than the modulus. In addition, the expression for `mul2` assumes that the input `m` is reduced modulo `n`, but this is not guaranteed since the `rem` is no longer taken after `ai+1`.