@b-mehta 2018-09-01
I've changed some things, removing some of the `rem` calls. When I benchmark, using `m * 2` instead of `shiftL` is noticeably faster, so I've kept it.  Since `Pollard` reduces to identical code as the tuples, I've left that as is also.  

The calls to `mod` in line 92 may not be strictly necessary - I originally kept them because `solveLinear'` was not exported and so I wanted to make sure the input was as expected. Also, it is likely that the `mod` call will reduce arguments into a smaller range, so the function may run faster, but I'm happy to remove them if necessary.