@Bodigrim 2018-09-01
`solveLinear'` should work fine for unreduced arguments. Even without reduction by modulo `bi-b2i` and `ai-b2i` are O(n) with probability close to 1, so there should not be any performance hit. I suggest to remove `mod` from line 92 for the sake of clarity.